+++ test_shell_app/Scripts/testapp1.py
@@ -1,7 +1,9 @@
-import io
-import sys
-from script import Script
-from shell import Shell
+import sys, os
+
+sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
+
+from test_shell_app.Utils.script import Script
+from test_shell_app.shell import Shell
 
 
 class TestApp1(Script):
